Moved Animation classes to namespace and folder
Only changed references and renamed Animation.cs to FrameAnimation.cs
