--- Assets/Scripts/Models/Animation.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 24f58493e14fa334594c761ea99c5170
-timeCreated: 1472207088
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
