--- Assets/Scripts/Models/CharacterAnimation.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: ef2796d7468cd294b942c10f00613117
-timeCreated: 1471934646
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
